+++ plugins/core/src/utils/fileDownload.ts
@@ -1,8 +1,12 @@
-// fileDownloads can be used to download the given data as file.
-// See: https://github.com/kennethjiang/js-file-download
+// fileDownload is a wrapper for the blodDownload function, but allows users to specify the data as string.
 export const fileDownload = (data: string, filename: string): void => {
   const blob = new Blob([data]);
+  blobDownload(blob, filename);
+};
 
+// blobDownload can be used to download the given data as file.
+// See: https://github.com/kennethjiang/js-file-download
+export const blobDownload = (blob: Blob, filename: string): void => {
   const blobURL =
     window.URL && window.URL.createObjectURL
       ? window.URL.createObjectURL(blob)